requirements.txt
Django==5.0.6
djangorestframework==3.15.2
Pillow==10.4.0

# Core ML deps (CPU works; CUDA speeds things up)
torch>=2.2.0
torchvision>=0.17.0

# Object detection (YOLOv8)
ultralytics==8.3.1

# Text-to-image generation (Stable Diffusion Turbo via diffusers)
diffusers==0.30.0
transformers==4.43.3
accelerate==0.33.0
safetensors==0.4.3
Django==5.0.6
djangorestframework==3.15.2
Pillow==10.4.0

# Core ML deps (CPU works; CUDA speeds things up)
torch>=2.2.0
torchvision>=0.17.0

# Object detection (YOLOv8)
ultralytics==8.3.1

# Text-to-image generation (Stable Diffusion Turbo via diffusers)
diffusers==0.30.0
transformers==4.43.3
accelerate==0.33.0
safetensors==0.4.3
